HardeningTest: move GosPackageState methods to a new TestUtils class
