+++ tests/HardeningTest/Android.bp
@@ -1,3 +1,8 @@
+filegroup {
+    name: "HardeningTestUtils",
+    srcs: ["src/grapheneos/hardeningtest/TestUtils.java"],
+}
+
 java_test_host {
     name: "HardeningTest",
     srcs: [
